refactor: Lotto에 필요한 상수 정보를 LottoApplication이 관리하도록 수정
- LottoConstant 제거
